requirements.txt
tensorflow==2.12.0
tensorflow-hub==0.16.1
keras==2.12.0
tensorflow==2.12.0
keras==2.12.0
numpy==1.23.5
pillow
streamlit
kagglehub
